fix(nc-gui): erd not rendering edges on initial mount
